Moved Hamcrest utilities to their own file
